Fix scrollToMakeVisible() so it can scroll to last cell and be used for keyboard navigation
The scrollToMakeVisible() method was updated so that when scrolling right, we calculate each columns' left and right positions and use them to check the far left column that we can scroll into, so our target column is still visible in the right. This method is heavier than the previous, but more robust when we have columns with different widths.
The remaining code from scrollToMakeVisible() was left untouched, because it's working great and it's faster than the approach above.
Tested with different column widths, number of fixed columns, different sizes for row numbers column and without this column.